+++ qqrobot-spring-boot-starter/src/main/java/de/honoka/qqrobot/starter/component/RobotConsole.java
@@ -1,10 +1,10 @@
 package de.honoka.qqrobot.starter.component;
 
-import de.honoka.sdk.util.system.ColorfulOutputStream;
+import cn.hutool.core.util.StrUtil;
+import de.honoka.sdk.util.gui.ColorfulOutputStream;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.SneakyThrows;
-import org.apache.commons.lang3.StringUtils;
 
 import java.awt.*;
 import java.io.PrintStream;
@@ -31,22 +31,18 @@ public class RobotConsole {
     @SneakyThrows
     private void changeSystemOut() {
         newOut = new ColorfulOutputStream(System.out, Color.LIGHT_GRAY);
-        newErr = new ColorfulOutputStream(System.err, new Color(
-                255, 107, 103));
+        newErr = new ColorfulOutputStream(System.err, new Color(255, 107, 103));
         newOut.setPrintMethod(bytes -> write(bytes, newOut));
         newErr.setPrintMethod(bytes -> write(bytes, newErr));
-        PrintStream newOutPrintStream = new PrintStream(newOut,
-                false, "UTF-8");
-        PrintStream newErrPrintStream = new PrintStream(newErr,
-                false, "UTF-8");
+        PrintStream newOutPrintStream = new PrintStream(newOut, false, StandardCharsets.UTF_8);
+        PrintStream newErrPrintStream = new PrintStream(newErr, false, StandardCharsets.UTF_8);
         System.setOut(newOutPrintStream);
         System.setErr(newErrPrintStream);
     }
 
     //两个输出流均使用此方法进行输出
     private synchronized void write(byte[] bytes, ColorfulOutputStream out) {
-        String str = new String(bytes, StandardCharsets.UTF_8)
-                .replace("\r", "");
+        String str = new String(bytes, StandardCharsets.UTF_8).replace("\r", "");
         if(!str.contains("\n")) {
             writeLine(str, out);
         } else {
@@ -63,10 +59,11 @@ public class RobotConsole {
             }
         }
         //检查保存的控制台输出是否超过最大行数
-        int lineCount = StringUtils.countMatches(lines, "\n") + 1;
+        int lineCount = StrUtil.count(lines, "\n") + 1;
         if(lineCount > maxLineCount) {
-            int newStartIndex = StringUtils.ordinalIndexOf(lines, "\n",
-                    lineCount - maxLineCount) + 1;
+            int newStartIndex = StrUtil.ordinalIndexOf(
+                lines, "\n", lineCount - maxLineCount
+            ) + 1;
             lines.delete(0, newStartIndex);
         }
     }
@@ -87,13 +84,14 @@ public class RobotConsole {
     }
 
     private void writeColorfulText(String str, Color color) {
-        if(StringUtils.isEmpty(str)) return;
+        if(StrUtil.isEmpty(str)) return;
         String template = "<pre style=\"color: %s;\">%s</pre>";
-        String colorStr = String.format("#%02X%02X%02X", color.getRed(),
-                color.getGreen(), color.getBlue());
+        String colorStr = String.format(
+            "#%02X%02X%02X", color.getRed(), color.getGreen(), color.getBlue()
+        );
         str = str.replace("<", "&lt;")
-                .replace(">", "&gt;")
-                .replace("\n", "<br>");
+            .replace(">", "&gt;")
+            .replace("\n", "<br>");
         lines.append(String.format(template, colorStr, str));
     }