+++ qqrobot-spring-boot-starter/src/main/java/de/honoka/qqrobot/starter/component/RobotConsoleWindow.java
@@ -3,7 +3,7 @@ package de.honoka.qqrobot.starter.component;
 import de.honoka.qqrobot.framework.api.Framework;
 import de.honoka.qqrobot.starter.RobotBasicProperties;
 import de.honoka.sdk.util.code.ThrowsRunnable;
-import de.honoka.sdk.util.system.gui.ConsoleWindow;
+import de.honoka.sdk.util.gui.ConsoleWindow;
 import de.honoka.sdk.util.text.TextUtils;
 import lombok.AccessLevel;
 import lombok.Getter;