+++ qqrobot-spring-boot-starter/web/tester/scripts/deploy.sh
@@ -15,5 +15,3 @@ npm run build:prod
 mv dist "$TARGET_DIR"
 cd "$TARGET_DIR"
 mv dist tester-framework
-
-git add tester-framework